I added a comment to the `getSegments` function to suggest that it could be refactored to use a more efficient algorithm
Hi team,

I added a comment to the `getSegments` function to suggest that it could be refactored to use a more efficient algorithm. The current implementation has a time complexity of O(n^2), which may be slow for large input sizes. I think this change could improve the performance of the function and make the code more scalable.

Let me know if you have any questions or concerns.

Best,
Benlad